move Options to common types
