+++ apps/browser/src/autofill/content/components/common-types.ts
@@ -1,7 +1,15 @@
+import { TemplateResult } from "lit";
+
 import { Theme } from "@bitwarden/common/platform/enums";
 
 export type IconProps = {
   color?: string;
   disabled?: boolean;
   theme: Theme;
+};
+
+export type Option = {
+  icon?: (props: IconProps) => TemplateResult;
+  text: string;
+  value: any;
 };
